requirements.txt
Django>=3.2,<4.0
openai
django-dotenv